+++ events/dialogController.js
@@ -5,23 +5,19 @@ const openDialog = async (e) => {
   const form = await fetch(`/components/${formType}`, {
     method: "GET",
   })
-    .then((response) => {
-      console.log(response);
-      return response.text();
+    .then((res) => {
+      console.log("res:", res);
+      return res.text();
     })
     .then((html) => {
       const parser = new DOMParser();
       const doc = parser.parseFromString(html, "text/html");
-      console.log("html:", html);
-      console.log("doc:", doc);
       return doc.querySelector("form");
     })
     .catch((err) => {
       console.log(err);
     });
 
-  // stopPropagation on the dialog element's child/children element(s)
-  form.addEventListener("click", stopPropagation);
   dialog.append(form);
   dialog.showModal();
 };
@@ -35,6 +31,7 @@ const onCloseHandler = () => {
   form.remove();
 };
 
+// stopPropagation on the dialog element's child/children element(s)
 const stopPropagation = (e) => {
   e.stopPropagation();
 };