+++ routes/componentsRouter.js
@@ -2,6 +2,8 @@ const { Router } = require("express");
 const {
   getUploadFileForm,
   getAddFolderForm,
+  postAddFolderForm,
+  postFilesUploadForm,
 } = require("../controllers/componentsController");
 
 const componentRouter = new Router();
@@ -9,5 +11,9 @@ const componentRouter = new Router();
 // GET requests
 componentRouter.get("/file", getUploadFileForm);
 componentRouter.get("/folder", getAddFolderForm);
+
+// POST requests
+componentRouter.post("/folder/create", postAddFolderForm);
+componentRouter.post("/files/upload", postFilesUploadForm);
 
 module.exports = componentRouter;