+++ routes/driveRouter.js
@@ -1,6 +1,6 @@
 const { Router } = require("express");
 const {
-  postFileUpload,
+  postFilesUpload,
   postFolderCreate,
 } = require("../controllers/driveController");
 
@@ -9,7 +9,7 @@ const driveRouter = new Router();
 // GET requests
 
 // POST requests
-driveRouter.post("/upload/file", postFileUpload);
+driveRouter.post("/upload/file", postFilesUpload);
 driveRouter.post("/folder/create", postFolderCreate);
 
 module.exports = driveRouter;